Switch to flat config format (#609)
This PR is effectively a complete rewrite of the repo, from an ESLint
plugin to a flat ESLint shared config.

Previously, we were exporting a plugin because of a limitation of
ESLint (https://github.com/eslint/eslint/issues/3458), which has since
been addressed by the flat config changes. In a nutshell, the old config
style required you to list all your config's plugins as peer
dependencies, and require your users to install them all. Since we have
eight dependencies, that would have been a burden, so we created our own
plugin that would re-export those plugins as if they were our own rules,
shifting the burden to our team to maintain.

Now, with the flat config, you can directly import dependencies, and
everything just works, so we now have only a single peer dependency,
ESLint itself. Much nicer.

In a perfect world, there would be no breaking changes. I've worked hard
to keep things as close to the previous version of the config as
possible. In practice, I had to update some of the dependencies to opt
into their flat configs, so there may be a handful of breaking changes
from those dependencies.